requirements.txt
# datasets
# decorator
Django
# google-ai-generativelanguage
# google-api-core
# google-auth
# google-auth-oauthlib
google-generativeai
# google-pasta
# googleapis-common-protos
happytransformer
httpcore
httpx
huggingface-hub
joblib
nltk
numpy
pandas
# pandasai==0.8.3
Pillow
psycopg2
python-decouple
python-dotenv
pytz
scikit-learn
scipy
sentence-transformers
sentencepiece
tensorflow
torch
transformers
tzdata
dj_database_url

